🐛 Fix message delete

+++ lib/handlers/commandHandler.ts
@@ -491,16 +491,28 @@ async function handleCommands(
                             });
                             break;
                         }
-                        const portalMessageId = helpers.getPortalMessageId(messageId)
+                        const portalMessageId =
+                            helpers.getPortalMessageId(messageId);
                         if (!portalMessageId) {
                             message.reply({
                                 content: "Message not found.",
                             });
                             break;
                         }
-                        const portalMessages = helpers.getPortalMessages(portalMessageId);
-                        for (const [messageId, portalMessage] of portalMessages) {
-                            const message = await helpers.safeFetchMessage({ messageId: portalMessage.messageId });
+                        const portalMessages =
+                            helpers.getPortalMessages(portalMessageId);
+                        for (const [
+                            messageId,
+                            portalMessage,
+                        ] of portalMessages) {
+                            const channel = await helpers.safeFetchChannel(
+                                portalMessage.channelId
+                            );
+                            if (!channel) continue;
+                            const message = await helpers.safeFetchMessage({
+                                messageId: portalMessage.messageId,
+                                channel
+                            });
                             if (!message) continue;
                             try {
                                 await message.delete();
